+++ src/ui-test/utils.ts
@@ -101,7 +101,7 @@ export async function selectContextMenuItem(command: string, menu: ContextMenu):
  * @param driver The WebDriver instance to use.
  * @param textArray An array of strings representing the texts to search for in the terminal view.
  * @param interval (Optional) The interval in milliseconds to wait between checks. Default is 2000ms.
- * @param timeout (Optional) The timout in milliseconds. Default is 60000ms.
+ * @param timeout (Optional) The timeout in milliseconds. Default is 60000ms.
  * @returns A Promise that resolves to a boolean indicating whether the terminal view has the texts or not.
  */
 export async function waitUntilTerminalHasText(driver: WebDriver, textArray: string[], interval = 2000, timeout = 60000): Promise<void> {