+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to the "vscode-debug-adapter-apache-camel" extension will be
 - Provide settings for specifying the Camel version to run integrations with Camel JBang CLI
 - Update default Camel version used for Camel JBang from 3.20.5 to 3.21.0
 - Upgrade Debug Adapter for Apache Camel to 0.8.0
+- Provide settings for specifying the Red Hat Maven Repository
+  - the repository is automatically used when the Camel version contains `redhat`
 
 ## 0.7.0